move textfield from login to components
also added ability to hide the label and custom label widths
