+++ source/views/settings/sections/login-credentials.js
@@ -1,7 +1,7 @@
 // @flow
 import React from 'react'
 import {Cell, Section} from 'react-native-tableview-simple'
-import {LoginField} from '../components/login-field'
+import {CellTextField} from '../../components/cells/textfield'
 import {LoginButton} from '../components/login-button'
 import {
   logInViaCredentials,
@@ -57,7 +57,7 @@ class CredentialsLoginSection extends React.Component {
         header="ST. OLAF LOGIN"
         footer="St. Olaf login enables the &quot;meals remaining&quot; feature."
       >
-        <LoginField
+        <CellTextField
           label="Username"
           _ref={ref => (this._usernameInput = ref)}
           disabled={loading}
@@ -69,7 +69,7 @@ class CredentialsLoginSection extends React.Component {
           value={username}
         />
 
-        <LoginField
+        <CellTextField
           label="Password"
           _ref={ref => (this._passwordInput = ref)}
           disabled={loading}
